+++ NEWS.md
@@ -1,3 +1,15 @@
+# anticlust 0.5.8
+
+* `anticlust` now includes the bicriterion algorithm for simultaneously 
+maximizing diversity and dispersion, proposed by Brusco et al.
+(<doi:10.1111/bmsp.12186>) and implemented by Martin Breuer (for details see [his bachelor thesis](https://www.cs.hhu.de/fileadmin/redaktion/Fakultaeten/Mathematisch-Naturwissenschaftliche_Fakultaet/Informatik/Algorithmische_Bioinformatik/Bachelor-_Masterarbeiten/2516084_ba_ifo_AbschlArbeit_klau_mapap102_mabre121_20200820_2320.pdf))
+  * It can be called from the main function `anticlustering()` by setting 
+  `method = "brusco"`; in this case only either dispersion or diversity is
+  maximized
+  * `bicriterion_anticlustering()` -- newly exported in this version -- can be used 
+  for a more fine grained usage of the Brusco et al. algorithm, fully using
+  its main functionality to optimize both dispersion as well as diversity
+
 # anticlust 0.5.7
 
 - Just an update to the documentation: Updating all references to the Papenberg & Klau paper after its "actual" publication in Psychological Methods.
@@ -41,7 +53,11 @@ minimizing differences with regard to means and variance is equally important.
 # anticlust 0.5.4-1
 
 - Internal change: `anticlustering()` with `objective = "dispersion"` now 
+<<<<<<< HEAD
 implements the local updating procedure [proposed by Martin Breuer](https://www.cs.hhu.de/fileadmin/redaktion/Fakultaeten/Mathematisch-Naturwissenschaftliche_Fakultaet/Informatik/Algorithmische_Bioinformatik/Bachelor-_Masterarbeiten/2516084_ba_ifo_AbschlArbeit_klau_mapap102_mabre121_20200820_2320.pdf). 
+=======
+implements the local updating procedure [proposed by Martin Breuer](https://www.cs.hhu.de/fileadmin/redaktion/Fakultaeten/Mathematisch-Naturwissenschaftliche_Fakultaet/Informatik/Algorithmische_Bioinformatik/Bachelor-_Masterarbeiten/2516084_ba_ifo_AbschlArbeit_klau_mapap102_mabre121_20200820_2320.pdf).
+>>>>>>> devel
 This leads to a considerable speedup when maximizing the dispersion, enabling the fast 
 processing of large data sets.